[LCR-36] refactor: 모듈 경로 수정
