+++ popi-api-gateway/src/main/java/com/lgcns/PopiApiGatewayApplication.java
@@ -1,4 +1,4 @@
-package com.lgcns.popiapigateway;
+package com.lgcns;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
