+++ popi-api-gateway/src/test/java/com/lgcns/PopiApiGatewayApplicationTests.java
@@ -1,4 +1,4 @@
-package com.lgcns.popiapigateway;
+package com.lgcns;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
